requirements.txt
python-telegram-bot==21.6
groq>=0.30.0
python-dotenv==1.0.1
aiofiles==24.1.0
pydub==0.25.1
openai>=1.61.0